Better wording for the show comments functionality.
User testing for S1 has shown that people miss the fact that they can show the comments column because they think that this is just an indicator how many comments are available. (Reasoning: It looks the same as other comment indicators, but it behaves different).

Thus the resolution: Make it look different too, to give people a reason to suspect that it does something. :)